requirements.txt
aiohttp==3.7.3
amqp==5.0.5
async-timeout==3.0.1
attrs==20.3.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
discord.py==1.6.0
idna==3.1
importlib-metadata==3.6.0
iniconfig==1.1.1
kombu==5.0.2
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.16
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
pytz==2021.1
redis==3.5.3
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.4.0
